Experiment with IDs for snapshots
The idea is that they are hashed of the place + stamp, to produce
shorter and therefor easier to type IDs, which also don't require the
place name to be passed.

4 characters for hex digits is *definitely* too short, what I'd rather
to is use a bigger alphabet to keep the char count low but that is going
to be a lot of work.