Change the way new tarsnap snapshots are made
First, make an intermediate one that includes an (extra) timestamp in
the name. Then, 'copy' it in to place by making it with the 'real' name.
This works around an awkward behavior in tarsnap that partials prevent
you from making a similarly-named new archive, and can't be copied.

Include cleanup for partial snapshots.

Still needs better messages and error handling, and a 'garbage
collection' mode.