Support time limits on backups
This sends tarsnap a SIQUIT after a specified time. The idea to to
start a backup at X AM and have it quit by Y AM, so that you can start
it again (roughly) where you picked up the next night.

Still fairly flawed, not the least of which is because it's per tarnsap
invocation, not per invocation of snappy.